refactor: rename leaderboards/utils.py to jsonl_io.py and update imports
